# docs/swagger.yaml
swagger: '2.0'
info:
  title: Sistema RH API
  version: 1.0.0
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
basePath: /api
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  Token:
    description: "este"
    type: apiKey
    name: Authorization
    in: header
paths:
  /users/login:
    post:
      summary: Existing user login
      description: Login for existing user
      tags:
        - User and Authentication
      operationId: Login
      parameters:
        - name: body
          in: body
          required: true
          description: Credentials to use
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
                format: password
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
  /users:
    post:
      summary: Register a new user
      description: Register a new user
      tags:
        - User and Authentication
      operationId: CreateUser
      parameters:
        - name: body
          in: body
          required: true
          description: Details of the new user to register
          schema:
            type: object
            properties:
              username:
                type: string
              email:
                type: string
              password:
                type: string
                format: password
            required:
              - username
              - email
              - password
      responses:
        '201':
          description: OK
        '422':
          description: Unexpected error
  /jobs:
    get:
      summary: Get all Jobs
      tags:
        - Jobs
      operationId: GetJobs
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
    post:
      summary: Create a Job
      description: Create a Job. Auth is required
      tags:
        - Jobs
      security:
        - Token: []
      operationId: CreateArticle
      parameters:
        - name: job
          in: body
          required: true
          description: Job to create
          schema:
            type: object
            properties:
              job:
                type: object
                properties:
                  title:
                    type: string
                  description:
                    type: string
                required:
                  - title
                  - description
            required:
              - job
      responses:
        '201':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
  /jobs/{id}:
    get:
      summary: Get a job
      description: Get a job. Auth is required
      tags:
        - Jobs
      operationId: GetJob
      parameters:
        - name: id
          in: path
          required: true
          description: Id of the job to get
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
    patch:
      summary: Update a job
      description: Update a job. Auth is required
      tags:
        - Jobs
      security:
        - Token: []
      operationId: UpdateJob
      parameters:
        - name: id
          in: path
          required: true
          description: Id of the job to get
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
    delete:
      summary: Delete a job
      description: Delete a job. Auth is required
      tags:
        - Jobs
      security:
        - Token: []
      operationId: DeleteJob
      parameters:
        - name: id
          in: path
          required: true
          description: Id of the job to get
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
  /jobs/{id}/applicants:
    get:
      summary: Get applicants for a job
      description: Get the applicants for a job. Auth is required
      tags:
        - Applicants
      operationId: GetJobApplicants
      parameters:
        - name: id
          in: path
          required: true
          description: Id of the job that you want to get applicants
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
    post:
      summary: Add a applicant to a job
      description: Add an applicant to a job. Auth is required
      tags:
        - Applicants
      security:
        - Token: []
      operationId: AddJobApplicant
      parameters:
        - name: id
          in: path
          required: true
          description: Id of the job
          type: string
        - name: applicant
          in: body
          required: true
          description: Applicant to add
          schema:
            type: object
            properties:
              applicant:
                type: object
                properties:
                  name:
                    type: string
                  email:
                    type: string
                  tel:
                    type: string
                  profile:
                    type: string
                required:
                  - name
                  - email
                  - tel
                  - profile
            required:
              - applicant
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
    patch:
      summary: Remove a applicant from a job
      description: Remove a applicant from a job. Auth is required
      tags:
        - Applicants
      security:
        - Token: []
      operationId: RemoveApplicant
      parameters:
        - name: id
          in: path
          required: true
          description: Id of the job to get
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
     
  /applicants/{id}:
    get:
      summary: Get an applicant
      description: Auth is required.
      tags:
        - Applicants
      operationId: GetApplicant.
      parameters:
        - name: id
          in: path
          required: true
          description: id applicant.
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error
    patch:
      summary: Update an Applicant
      description: Update a Applicant. Auth is required
      tags:
        - Applicants
      security:
        - Token: []
      operationId: UpdateApplicant
      parameters:
        - name: id
          in: path
          required: true
          description: id of the job to get
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error 
    delete:
      summary: Delete a job
      description: Delete a job. Auth is required
      tags:
        - Applicants
      security:
        - Token: []
      operationId: DeleteApplicant
      parameters:
        - name: id
          in: path
          required: true
          description: id of the job to get
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '422':
          description: Unexpected error     
    